First draft of repository pattern implementation
