+++ GirafRest/Repositories/GirafUserRepository.cs
@@ -1,8 +1,10 @@
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.EntityFrameworkCore;
 using GirafRest.Models;
 using GirafRest.IRepositories;
 using GirafRest.Data;
+using System.Threading.Tasks;
 
 namespace GirafRest.Repositories
 {
@@ -11,5 +13,24 @@ namespace GirafRest.Repositories
         public GirafUserRepository(GirafDbContext context) : base(context)
         {
         }
+
+        /// <summary>
+        /// Method for loading user from context and eager loading fields requied to read their <b>week schedules</b>
+        /// </summary>
+        /// <param name="id">id of user to load.</param>
+        /// <returns>A <see cref="GirafUser"/> with <b>all</b> related data.</returns>
+        public async Task<GirafUser> LoadUserWithWeekSchedules(string id)
+            => await Context.Users
+                //First load the user from the database
+                .Where(u => u.Id.ToLower() == id.ToLower())
+                // then load his week schedule
+                .Include(u => u.WeekSchedule)
+                .ThenInclude(w => w.Thumbnail)
+                .Include(u => u.WeekSchedule)
+                .ThenInclude(w => w.Weekdays)
+                .ThenInclude(wd => wd.Activities)
+                .ThenInclude(e => e.Pictograms)
+                //And return it
+                .FirstOrDefaultAsync();
     }
 }
